# src/Hercules.ED.Enrichment/ServicioAPI/openapi.json
{
    "info": {
        "description": "This module classifies papers, protocols or open-source code projects accoding to a predefined thematic descriptor taxonomy and free specific descriptors.",
        "title": "Hercules REST API for descriptors",
        "version": "1.0"
    },
    "paths": {
        "/thematic": {
            "post": {
                "tags": [
                    "Hercules"
                ],
                "description": "Hercules thematic keyword API.",
                "parameters": [
                    {
                        "in": "body",
                        "required": false,
                        "name": "body",
                        "schema": {
                            "$ref": "#/components/schemas/ThematicRequest"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "schema": {
                            "$ref": "#/components/schemas/ThematicResponse"
                        },
                        "description": "returns the request object + a new property called topics, which contains a dictionary with the tags found and theis probabilities"
                    }
                }
            }
        },
        "/specific": {
            "post": {
                "tags": [
                    "Hercules, Specific keywords"
                ],
                "description": "Hercules Specific keyword API endpoint.",
                "parameters": [
                    {
                        "in": "body",
                        "required": false,
                        "name": "body",
                        "schema": {
                            "$ref": "#/components/schemas/SpecificRequest"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "schema": {
                            "$ref": "#/components/schemas/SpecificResponse"
                        },
                        "description": "returns the request object + a new property called topics, which contains a dictionary with the tags found and theis probabilities"
                    }
                }
            }
        }
    },
    "openapi": "3.0.0",
    "components": {
        "schemas": {
            "ThematicRequest": {
                "type": "object",
                "properties": {
                    "text": {
                        "type": "string",
                        "description": "Input text to classify"
                    },
                    "author_affiliation": {
                        "type": "string",
                        "description": "Author(s)'s affilliation(s)"
                    },
                    "pdf_url": {
                        "type": "string",
                        "description": "URL of the pdf version of the paper. If pdf_url is given no other parameter is required, except for the rotype"
                    },
                    "abstract": {
                        "type": "string",
                        "description": "Description/abstract of the Input article to classify"
                    },
                    "author_name": {
                        "type": "string",
                        "description": "Author(s) of the paper"
                    },
                    "title": {
                        "type": "string",
                        "description": "Title of the input article to classify"
                    },
                    "rotype": {
                        "type": "string",
                        "description": "Type of the input text, currently available: 'bio-protocol'=protocol descriptions | 'sourceForge'=software descriptions | 'papers'=scientific papers."
                    },
                    "journal": {
                        "type": "string",
                        "description": "Journal where the paper was published"
                    }
                },
                "required": [
                    "rotype"
                ]
            },
            "ThematicResponse": {
                "type": "object",
                "properties": {
                    "text": {
                        "type": "string",
                        "description": "Input text to classify"
                    },
                    "author_affiliation": {
                        "type": "string",
                        "description": "Author(s)'s affilliation(s)"
                    },
                    "topics": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        }
                    },
                    "pdf_url": {
                        "type": "string",
                        "description": "URL of the pdf version of the paper. If pdf_url is given no other parameter is required, except for the rotype"
                    },
                    "abstract": {
                        "type": "string",
                        "description": "Description/abstract of the Input article to classify"
                    },
                    "author_name": {
                        "type": "string",
                        "description": "Author(s) of the paper"
                    },
                    "title": {
                        "type": "string",
                        "description": "Title of the input article to classify"
                    },
                    "rotype": {
                        "type": "string",
                        "description": "Type of the input text, currently available: 'bio-protocol'=protocol descriptions | 'sourceForge'=software descriptions."
                    },
                    "journal": {
                        "type": "string",
                        "description": "Journal where the paper was published"
                    }
                },
                "required": [
                    "rotype"
                ]
            },
            "SpecificRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "Title of the input article to classify"
                    },
                    "body": {
                        "type": "string",
                        "description": "Main text of the paper. If no body is present the systems resorts to short texts keyphrase extractor"
                    },
                    "rotype": {
                        "type": "string",
                        "description": "Type of the input text, currently available: 'bio-protocol'=protocol descriptions | 'sourceForge'=software descriptions."
                    },
                    "abstract": {
                        "type": "string",
                        "description": "Abstract of the Input article to classify"
                    }
                },
                "required": [
                    "abstract",
                    "rotype"
                ]
            },
            "SpecificResponse": {
                "type": "object",
                "properties": {
                    "body": {
                        "type": "string",
                        "description": "Main text of the paper. If no body is present the systems resorts to short texts keyphrase extractor"
                    },
                    "topics": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        }
                    },
                    "abstract": {
                        "type": "string",
                        "description": "Abstract of the Input article to classify"
                    },
                    "title": {
                        "type": "string",
                        "description": "Title of the input article to classify"
                    },
                    "rotype": {
                        "type": "string",
                        "description": "Type of the input text, currently available: 'bio-protocol'=protocol descriptions | 'sourceForge'=software descriptions."
                    }
                },
                "required": [
                    "abstract",
                    "rotype"
                ]
            }
        }
    }
}